refactor: fold sacc_interop into sacc_io — one SACC module
Cail's ruling: a single sacc_io.py carries the format contract and the
external-tool converters (CosmoSIS 2pt-FITS, OneCovariance). Public API
unchanged; the sacc_io import name every downstream branch uses is untouched.
